requirements.txt
pillow>=11.0.0
numpy>=2.1.0
opencv-python>=4.8.0
flask>=3.0.0
matplotlib>=3.8.0
